move component and remove comments
